data-onchain-ingestor/pyproject.toml
[tool.poetry]
name = "data-onchain-ingestor"
version = "0.1.0"
description = ""
authors = ["HQ platform <giang@hq.xyz>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
hypersync = "^0.7.14"
asyncio = "^4.0.0"
pydantic = "^2.7.4"
colorlog = "^6.8.2"
deltalake = "^1.1.4"
pyarrow = "^21.0.0"
polars = {extras = ["deltalake"], version = "^1.33.1"}
duckdb = "^1.0.0"
web3 = "^6.20.4"
xxhash = "^3.4.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
psycopg2 = "^2.9.9"
pandas = "^2.2.2"
redis = "^6.4.0"
types-redis = "^4.6.0.20240425"
temporalio = "^1.6.0"
slack-sdk = "^3.31.0"
sentry-sdk = "^2.10.0"
adbc-driver-postgresql = "^1.1.0"
fastapi = {extras = ["standard"], version = "^0.118.0"}
pydantic-settings = "^2.11.0"
asyncpg = "^0.30.0"
uvicorn = {extras = ["standard"], version = "^0.37.0"}


[tool.poetry.group.dev.dependencies]
setuptools = "^70.0.0"
ruff = "^0.4.9"
mypy = "^1.10.0"
pre-commit = "^3.7.1"
pytest = "^8.2.2"
isort = "^5.13.2"
colorlog = "^6.8.2"


[tool.poetry.group.test.dependencies]
pytest-mock = "^3.14.0"
pytest-postgresql = "^6.0.0"
pytest-redis = "^3.1.2"
fakeredis = "^2.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
follow_imports = "silent"
python_version = "3.12"
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
no_implicit_optional = true
strict = true
ignore_missing_imports = true
disable_error_code = ["call-overload", "misc" , "assignment", "union-attr", "call-arg", "arg-type"]
